Add support for passing the domain search list to LWIP
It is very common on corporate proxies to resolve host names
with a no dot hostname in conjunction with a domain search
list. This adds support for passing the search list to LWIP.

Note that for this to function, the request needs to make it to
the LWIP socks proxy. This usually means adding a:

if (isPlainHostName(host))
	return "SOCKS5 127.0.0.1:1080";

or similar to your proxy.pac.